Renames /classes to /models; updates imports
