+++ src/models/player.py
@@ -1,5 +1,6 @@
 from .item import Item
 
+
 class Player:
 
   def __init__(self, name, inv=None):
